chore(integration): Rename error to errors and fix "The inferred type of X cannot be named without a reference to Y"
TS issue: https://github.com/microsoft/TypeScript/issues/47663
