+++ packages/nextjs/src/client-boundary/hooks.ts
@@ -13,4 +13,4 @@ export {
   useUser,
 } from '@clerk/clerk-react';
 
-export { isClerkAPIResponseError, isEmailLinkError, isKnownError, isMetamaskError } from '@clerk/clerk-react/error';
+export { isClerkAPIResponseError, isEmailLinkError, isKnownError, isMetamaskError } from '@clerk/clerk-react/errors';
