+++ packages/react/subpaths.mjs
@@ -1,7 +1,10 @@
+import packageJson from './package.json' assert { type: 'json' };
 // This file is a helper for the "subpath-workaround.mjs" script
 // We have to polyfill our "exports" subpaths :cry:
 
-export const subpathNames = ['internal', 'error'];
+export const subpathNames = Object.keys(packageJson.exports)
+  .filter(k => k.startsWith('./') && k !== './' && k !== './package.json')
+  .map(k => k.replace('./', ''));
 
 export const subpathFoldersBarrel = [];
